Move embeddedHTML functions outside of index
