+++ packages/server/src/__tests__/plugin/landingPage/getEmbeddedSandboxHTML.test.ts
@@ -1,5 +1,5 @@
-import { getEmbeddedSandboxHTML } from '../default';
-import type { LandingPageConfig } from '../default/types';
+import { getEmbeddedSandboxHTML } from '../../../plugin/landingPage/default/getEmbeddedHTML';
+import type { LandingPageConfig } from '../../../plugin/landingPage/default/types';
 
 const version = '_latest';
 expect.addSnapshotSerializer(require('jest-serializer-html'));
